+++ avista_sensors/__init__.py
@@ -1,6 +1,7 @@
 """This module is used to provide the capability to read data from sensors attached to the IoT device"""
 import avista_sensors.impl
-import avista_sensors.manager_state
+import avista_sensors.sweep_state
 import avista_sensors.processor_loader
 import avista_sensors.sensor_sweep
 import avista_sensors.sensor_processor
+import avista_sensors.data_transporter